+++ packages/client/src/cache.ts
@@ -569,6 +569,7 @@ export const exchange = cacheExchange({
     ReserveStatus: () => null,
     ReserveSummary: () => null,
     ReserveUserState: () => null,
+    SharedBalanceNativeAsset: () => null,
     SpokePositionManager: () => null,
     SpokeUserPositionManager: () => null,
     StableVaultPendingAvailability: () => null,
@@ -609,5 +610,6 @@ export const exchange = cacheExchange({
     UserSupplyItem: () => null,
     WithdrawSwap: () => null,
     WithdrawSwapQuoteResult: () => null,
+    WrappedNativeAsset: () => null,
   },
 });